Oh Shoot!.. added shooting
Added shooting and the ability to aim in a cone.
Look in the TestShooting scene to see it hitting an enemy.
You shoot with "A" on a controller.
The aim arrow and cone can have their mesh renderers turned off and
still function as expected.

There's a player and a bullet prefab that gets spawned in.  You can drop
this into any scene.