Removed Prototype Assets and Project
I removed the prototype assets and project.  Don't worry though, I've
backed them up in case we need to refer back to something or if we want
to use any of the models from the prototype.  However, given the new art
style, I'm not sure if any of the models fit in anymore.