I did some stuff.
I setup a sample level.
The level contains 1 player, and that's all that's supported right now.
Implemented a checkpoint system which tracks whether the player took a
valid path to the finish or not.  I also added a timer which tracks your
overall time, your last laps time, and your current laps time.
The current lap is displayed i.e "Lap: 1/3".
There's a countdown at the beginning of the level.  You can't start the
race until it's done, and the timers won't start until then either.
When you finish the race, you start to slow down, and you're told you
finished.  The overall time, last, and current, should all be displayed.
Last and Current should be the same.  Haven't verified though.
I added in some fonts I found online, the one I have currently on fits
the theme best in my opinion.  We can look for others and swap them out
if we want.
The materials in the scene aren't too reflective.  It looks more like a
90s cartoon that way with less reflection and more of a flat color (like
a toon shader almost).  Also can be toyed with.
Maybe missed some other stuff.. Not sure