NEARLY finished implementing screens.
There's probably a bug or two I missed, and I'm also very tired and
ready for bed.  I'll look into them tomorrow.
Also, pausing the game pauses everything except the movement of the
player.  I need to reach into the rigidbody and save some variables upon
pausing, then set them to 0.  When unpaused, they are restored and the
player moves as normal.  There's probably a better way to do this, but
as it's setup I think it's the best/easiest way.