Multiplayer Update
The game now supports 1-4 players based on a variable that you can
change in the Race Manager object's Player Setup script. Each player has
their own HUD and the game keeps track of their laps separately. I wasnt
able to interface the character selection screen with the multiplayer
aspects but I'll work with Trevor on that.