Control Manager Update 2
The control manager is online! There are four files in Assets/Controls
that you can modify to customize the control setup. Right now it is
defaulted to use a keyboard input with A/D as pedals and
LeftArrow/RightArrow as steering. You can of course change this to
whatever you'd like it to be.