Bat Ultimate Update
The bat is implemented as an ultimate. Hit Y once you have adequate
charge and it appears and smashes in front of you. Charge is gained by
pedalling and by people running into your marshmallows. Also made
marshmallows a bit bigger. Right now nothing happens when you use the
bat other than the shitty animation and there isn't a charge meter but
that will be in the works.