Bat Hits and more Place Finding fixes
Place finding now works finally after all this time, and the display
shows what place you are in. If you swing your bat, everyone in front of
you pops up on collision with the ground. For a good time, set the
charge needed to ulti on the weapon handler to 0 and then mash the Y
button to send people flying.